add claim queue message action and refactor publish message action to remove circular dependency introduced with claim queue message
