+++ server/services/publish-message-action.js
@@ -1,0 +1,19 @@
+const MessageSender = require('./messaging/message-sender')
+const config = require('../config')
+
+const calculationSender = new MessageSender(config.calculationQueueConfig, config.calculationQueueConfig.queueUrl)
+const scheduleSender = new MessageSender(config.scheduleQueueConfig, config.scheduleQueueConfig.queueUrl)
+
+async function publishMessageAction (claim) {
+  try {
+    await Promise.all([
+      calculationSender.sendMessage(claim),
+      scheduleSender.sendMessage(claim)
+    ])
+  } catch (err) {
+    console.log(err)
+    throw err
+  }
+}
+
+module.exports = { publishMessageAction }
